+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added ⭐
+
+* You can now check if a `TextEdit` lost keyboard focus with `response.lost_kb_focus`.
+
 ### Changed 🔧
 
 * Pressing enter in a single-line `TextEdit` will now surrender keyboard focus for it